Fix VS2003 cast
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@9874 6f19259b-4bc3-4df7-8a09-765794883524

+++ EmbeddedPkg/EmbeddedMonotonicCounter/EmbeddedMonotonicCounter.c
@@ -49,7 +49,7 @@ GetNextHighMonotonicCount (
   
   gCurrentMonotonicCount += 0x0000000100000000ULL;
   
-  *HighCount = RShiftU64 (gCurrentMonotonicCount, 32) & 0xFFFFFFFF;
+  *HighCount = (UINT32)RShiftU64 (gCurrentMonotonicCount, 32) & 0xFFFFFFFF;
 
   return EFI_SUCCESS;
 }
